+++ modules/category/infras/controller/http-gin/update_category_by_id_api.go
@@ -1,0 +1,34 @@
+package httpgin
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
+	"github.com/ntttrang/go-food-delivery-backend-service/modules/category/service"
+	"github.com/ntttrang/go-food-delivery-backend-service/shared/datatype"
+)
+
+func (ctrl *CategoryHttpController) UpdateCategoryByIdAPI(c *gin.Context) {
+	id, err := uuid.Parse(c.Param("id"))
+
+	if err != nil {
+		panic(datatype.ErrBadRequest.WithError(err.Error()))
+	}
+
+	var req service.CategoryUpdateReq
+	if err := c.ShouldBindJSON(&req); err != nil {
+		panic(datatype.ErrBadRequest.WithError(err.Error()))
+	}
+	req.Id = id
+
+	// Get requester from context for authorization
+	requester := c.MustGet(datatype.KeyRequester).(datatype.Requester)
+	req.Requester = requester
+
+	if err := ctrl.updateByIdCommandHandler.Execute(c.Request.Context(), req); err != nil {
+		panic(err)
+	}
+
+	c.JSON(http.StatusOK, gin.H{"data": gin.H{"id": id}})
+}