+++ CHANGELOG.md
@@ -6,3 +6,8 @@
 
 ## 2025-06-25 v.0.0.2
 - Implement media module
+
+## 2025-06-25 v0.0.3
+- LCRUD category
+- Proto file
+- GRPC (server) FindByIds